Modificação da tela Home e Análise de clubes
